requirements.txt
# Core dependencies for r-u.live DNS management
PyYAML>=6.0
requests>=2.28.0
jsonschema>=4.17.0

# Optional dependencies for enhanced features
dnspython>=2.3.0
matplotlib>=3.6.0
seaborn>=0.12.0
pandas>=1.5.0

# Development dependencies (optional)
pytest>=7.2.0
black>=22.10.0
flake8>=5.0.0
mypy>=0.991
